리팩토링 : 불필요한 컴포넌트화 분리
