+++ src/pages/Items.jsx
@@ -1,9 +1,11 @@
-import ProductManagement from "../components/Items/ProductManagement";
 import useProductsAll from "../hooks/useProductsAll";
 import ProductsList from "../components/Items/ProductsList";
 import PageNation from "../components/common/PageNation";
 import "./Items.scss";
 import useProductsFavorite from "../hooks/useProductsFavorite";
+import SearchInput from "../components/common/SearchInput";
+import Button from "../components/common/Button";
+import DropDown from "../components/common/DropDown";
 
 function Items() {
   const {
@@ -25,6 +27,18 @@ function Items() {
   const isEmpty =
     allItems.length === 0 && !searchError && !loadingError && !isLoading;
 
+  const options = [
+    { label: "최신순", value: "recent" },
+    { label: "좋아요순", value: "favorite" },
+  ];
+  const handleSelect = (value) => {
+    if (value === "recent") {
+      handleNewestClick();
+    } else if (value === "favorite") {
+      handleBestClick();
+    }
+  };
+
   return (
     <main className='page-items'>
       <div className='container'>
@@ -38,11 +52,16 @@ function Items() {
         <div className='product-area all'>
           <div className='product-category'>
             <h2 className='product-category-name'>전체 상품</h2>
-            <ProductManagement
-              onSubmit={handleSearchSubmit}
-              onBest={handleBestClick}
-              onNewest={handleNewestClick}
-            />
+            <SearchInput onSubmit={handleSearchSubmit} />
+            <Button
+              link={true}
+              href='/additem'
+              className='additem'
+              styleType='square blue'
+            >
+              상품 등록하기
+            </Button>
+            <DropDown options={options} onSelect={handleSelect} />
           </div>
           {isLoading && <p>로딩 중 입니다...</p>}
           <ProductsList list={allItems} imageSize='middle' />
